requirements.txt
django-restframework
django-rest-knox